+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Minimax exercise".
 - Tests for "Constrained latin squares".
 - Tests for "Crossword".
+- Tests for "Futoshiki solver".
 
 ## [1.11.0] - 2022-07-31
 ### Added